Half the lies they tell about me aren't true

+++ 0x05-pointers_arrays_strings/6-puts2.c
@@ -9,7 +9,7 @@ void puts2(char *str)
 {
     while (*str)
     {
-        putchar((int)*str);
+        putchar(*str);
         str = str + 2;
     }
     putchar('\n');
